requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-lru==2.0.4
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
dnspython==2.4.2
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
motor==3.3.2
multidict==6.0.4
nltk==3.8.1
oauthlib==3.2.2
openai==1.3.4
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.6.0
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
tweepy==4.14.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.3




